rename hierarchy imports in worker.eviction
