+++ core/src/main/java/tachyon/worker/eviction/EvictPartialLRU.java
@@ -21,8 +21,8 @@ import java.util.List;
 import java.util.Set;
 
 import tachyon.Pair;
-import tachyon.worker.hierarchy.BlockInfo;
-import tachyon.worker.hierarchy.StorageDir;
+import tachyon.worker.tiered.BlockInfo;
+import tachyon.worker.tiered.StorageDir;
 
 /**
  * Used to evict blocks in certain StorageDir by LRU strategy.
